rename voter_registration mods to cip36_registration
