+++ catalyst-gateway/bin/src/cardano/cip36_registration/61284.cddl
