+++ catalyst-gateway/bin/src/cardano/cip36_registration/61285.cddl
