+++ catalyst-gateway/bin/src/cardano/cip36_registration/cip36.cddl
