+++ catalyst-gateway/bin/src/cardano/cip36_registration/mod.rs
