+++ catalyst-gateway/bin/src/event_db/cardano/cip36_registration/insert_cip36_registration.sql
