+++ catalyst-gateway/bin/src/event_db/cardano/cip36_registration/select_cip36_registration.sql
