+++ catalyst-gateway/bin/src/event_db/cardano/mod.rs
@@ -3,4 +3,4 @@
 pub(crate) mod config;
 pub(crate) mod chain_state;
 pub(crate) mod utxo;
-pub(crate) mod voter_registration;
+pub(crate) mod cip36_registration;
